backend/requirements.txt

# Fastening light Python Framework for API build
fastapi == 0.65.1

# Fast ASGI server implementation which helps in the interaction between your application and the backend 
uvicorn == 0.14.0

# Motor is a full-featured, non-blocking IO MongoDB driver for Python 
motor == 2.4.0

# Pydantic is a library for data validation and settings management based on Python type hinting
pydantic == 1.8.2

# Google
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-language==2.12.0
googleapis-common-protos==1.62.0
oauthlib==3.2.2

uvicorn==0.25.0

beautifulsoup4==4.12.3